requirements.txt
# eSIM Manager System Dependencies
# Production-ready requirements for GSMA-compliant platform

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1

# Caching and Message Queue
redis==5.0.1
celery==5.3.4
kombu==5.3.4

# Security and Cryptography
cryptography==41.0.7
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
aiohttp==3.9.1
httpx==0.25.2

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
mkdocs-swagger-ui-tag==0.6.8

# Deployment
gunicorn==21.2.0
docker==6.1.3
kubernetes==28.1.0

# GSMA Standards Support
asn1crypto==1.5.1
pyasn1==0.5.1
pyasn1-modules==0.3.0

# Performance
orjson==3.9.10
ujson==5.8.0

# Utilities
click==8.1.7
rich==13.7.0
typer==0.9.0